ml_logic/requirements.txt
django
djangorestframework
tensorflow
scikit-learn
pillow
numpy
